类型修改 Integer -> Long

+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/game/entity/GameActivity.java
@@ -30,7 +30,7 @@ public class GameActivity implements Serializable {
      * 活动id
      */
     @TableId(type = IdType.AUTO)
-    private java.lang.Integer id;
+    private java.lang.Long id;
 
     /**
      * 活动唯一标识, e.g. "buff", "worldboss"
